[#2649] Move `safe_next_or_default_url` to workshops.utils
