+++ amy/autoemails/tests/test_utils.py
@@ -6,11 +6,7 @@ from rq.worker import SimpleWorker
 from rq_scheduler.utils import to_unix
 
 from autoemails.tests.base import FakeRedisTestCaseMixin, dummy_fail_job, dummy_job
-from autoemails.utils import (
-    check_status,
-    safe_next_or_default_url,
-    scheduled_execution_time,
-)
+from autoemails.utils import check_status, scheduled_execution_time
 
 
 class TestScheduledExecutionTime(FakeRedisTestCaseMixin, TestCase):
@@ -169,41 +165,3 @@ class TestCheckStatus(FakeRedisTestCaseMixin, TestCase):
         # test
         rv = check_status(job, self.scheduler)
         self.assertEqual(rv, "deferred")
-
-
-class TestSafeNextOrDefaultURL(TestCase):
-    def test_default_url_if_next_empty(self):
-        # Arrange
-        next_url = None
-        default_url = "/dashboard/"
-        # Act
-        url = safe_next_or_default_url(next_url, default_url)
-        # Assert
-        self.assertEqual(url, default_url)
-
-    def test_default_url_if_next_not_provided(self):
-        # Arrange
-        next_url = ""
-        default_url = "/dashboard/"
-        # Act
-        url = safe_next_or_default_url(next_url, default_url)
-        # Assert
-        self.assertEqual(url, default_url)
-
-    def test_default_url_if_next_not_safe(self):
-        # Arrange
-        next_url = "https://google.com"
-        default_url = "/dashboard/"
-        # Act
-        url = safe_next_or_default_url(next_url, default_url)
-        # Assert
-        self.assertEqual(url, default_url)
-
-    def test_next_url_if_next_safe(self):
-        # Arrange
-        next_url = "/admin/"
-        default_url = "/dashboard/"
-        # Act
-        url = safe_next_or_default_url(next_url, default_url)
-        # Assert
-        self.assertEqual(url, next_url)
